+++ src/main/java/com/simpllyrun/srcservice/api/feed/domain/PostLike.java
@@ -1,4 +1,4 @@
-package com.simpllyrun.srcservice.api.domain.feed;
+package com.simpllyrun.srcservice.api.feed.domain;
 
 import com.simpllyrun.srcservice.api.domain.BaseDomain;
 import com.simpllyrun.srcservice.api.domain.User;
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @Entity
 @NoArgsConstructor
-@Table(name = "comment")
-public class Comment extends BaseDomain {
+@Table(name = "post")
+public class PostLike extends BaseDomain {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -19,16 +19,14 @@ public class Comment extends BaseDomain {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
+
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id")
     private Post post;
-    private String content;
-
 
     @Builder
-    public Comment(User user, Post post, String content) {
+    public PostLike(User user, Post post) {
         this.user = user;
         this.post = post;
-        this.content = content;
     }
 }
