+++ src/main/java/com/simpllyrun/srcservice/api/feed/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.simpllyrun.srcservice.api.repository.feed;
+package com.simpllyrun.srcservice.api.feed.repository;
 
-import com.simpllyrun.srcservice.api.domain.feed.Post;
+import com.simpllyrun.srcservice.api.feed.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
